Concerns about the validity of PAI across devices
Concerns about the validity of PAI estimates based on accuracy and differences in measurements in the reference equipment vs. the currently used equipment, and between different instances of the latter (smart watches, etc.).

It ought to be mentioned somewhere by someone capable to assess it intellectually. Looks like I'm the one for now...